+++ package/src/routelar/generation/generation.utils.ts
@@ -6,19 +6,3 @@ export const excludeRoot = routes =>
 
     return acc;
   }, {});
-
-export function flatRoutes(routes) {
-  if (routes.root === undefined) {
-    return routes;
-  }
-
-  const flattenRoutes = flatRoutes(routes.root);
-
-  const normalizedRoutes =
-    Object.keys(routes.root).length > 0 ? excludeRoot(routes) : routes;
-
-  return {
-    ...normalizedRoutes,
-    ...flattenRoutes
-  };
-}
